+++ src/IDBridge/Public/Source/Get-SourceDataGSheet.ps1
@@ -71,7 +71,6 @@ function Get-SourceDataGSheet {
         "PersonType"
         "JobTitle"
         "TerminationDate"
-        "Word"
         "Process"
     )